+++ lib/logstash-logger/logger.rb
@@ -1,12 +1,13 @@
 require 'logger'
 
 class LogStashLogger < ::Logger
-  DEFAULT_CONNECTION_TYPE = :udp
+  include ::LogStash::TaggedLogging
 
   attr_reader :connection
 
-  def initialize(host, port, type = DEFAULT_CONNECTION_TYPE)
-    @connection = ::LogStash::Connection.new(host, port, type)
+  def initialize(*args)
+    connection_options = extract_connection_opts(*args)
+    @connection = ::LogStash::Connection.new(connection_options)
     super(@connection)
     self.formatter = Formatter.new
   end
@@ -15,5 +16,18 @@ class LogStashLogger < ::Logger
     !!@connection.flush
   end
 
-  include ::LogStash::TaggedLogging
+  protected
+
+  def extract_connection_opts(*args)
+    if args.length > 1
+        puts "[LogStashLogger] (host, port, type) constructor is deprecated. Please use an options hash instead."
+        host, port, type = *args
+        {host: host, port: port, type: type}
+      elsif Hash === args[0]
+        args[0]
+      else
+        fail ArgumentError, "Invalid LogStashLogger options"
+    end
+  end
+
 end