+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 0.5.0 (Unreleased)
 - Support for tagged logging. The interface was extracted from `ActiveSupport::TaggedLogging`
 and outputs to the `tags` key.
+- The `(host, port, type)` constructor has been deprecated in favor of an options hash constructor.
 - Support for Rails configuration.
+- Output to STDOUT in Rails console.
+- `host` is no longer required. It will default to `0.0.0.0`, the same default port that logstash listens on.
 - Changed `source` to `host` to match what the latest logstash expects.
 - Output event timestamp consistently even if `Time#to_json` is overridden.
 - Refactoring.